refactor(linter): merge vitest/valid-title into jest/valid-title
Consolidates the `valid-title` rule implementation for both Jest and Vitest
into a single rule located in the jest directory. This unifies the logic
and reduces code duplication while maintaining full compatibility with both
testing frameworks.

Changes:
- Merged vitest configuration options into jest/valid-title:
  - Added `allowArguments` option (allows identifier arguments)
  - Preserved all existing jest options
- Merged vitest-specific logic:
  - Skip `.extend()` calls (test.extend(), it.extend())
  - Respect `ctx.settings().vitest.typecheck` for describe blocks
  - Support identifier arguments when `allowArguments` is enabled
- Enhanced pattern compilation:
  - Handle JS regex literals (e.g., /pattern/flags)
  - Strip `/` prefix and suffix from patterns automatically
  - Backward compatible with existing pattern formats
- Improved error messages:
  - Empty title errors now include function name
  - Example: `\`describe\` should not have an empty title`
- Updated documentation to reference both Jest and Vitest
- Merged all vitest test cases into jest test suite
- Removed vitest/valid-title.rs and module registration
- Updated snapshots with improved error messages

All 794 tests passing. The rule now supports both frameworks from a single
location while maintaining backward compatibility with existing configs.
