+++ crates/oxc_linter/src/rules.rs
@@ -637,7 +637,6 @@ pub(crate) mod vitest {
     pub mod prefer_to_be_object;
     pub mod prefer_to_be_truthy;
     pub mod require_local_test_context_for_concurrent_snapshots;
-    pub mod valid_title;
 }
 
 pub(crate) mod node {
@@ -1259,7 +1258,6 @@ oxc_macros::declare_all_lint_rules! {
     vitest::prefer_to_be_object,
     vitest::prefer_to_be_truthy,
     vitest::require_local_test_context_for_concurrent_snapshots,
-    vitest::valid_title,
     vue::define_emits_declaration,
     vue::define_props_declaration,
     vue::define_props_destructuring,
